Move task manager client and restruct task manager pkg.
